fc-detection-project/requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Data manipulation
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging and monitoring
tqdm>=4.65.0
tensorboard>=2.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
